requirements.txt
asgiref==3.5.2
astroid==2.12.12
autopep8==2.0.0
black==22.10.0
click==8.1.3
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
Django==3.2.16
django-cors-headers==3.13.0
django-enumfields==2.1.1
django-extensions==3.2.1
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
generics==3.9.0
gunicorn==20.1.0
lazy-object-proxy==1.8.0
MarkupPy==1.14
mccabe==0.7.0
mypy-extensions==0.4.3
odfpy==1.4.1
openpyxl==3.0.10
pathspec==0.10.1
pep8==1.7.1
platformdirs==2.5.2
pycodestyle==2.9.1
pylint==2.15.5
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.0.0
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
unicodecsv==0.14.1
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
APScheduler==3.0.0